Create README.md
This is a Final Project submission for my course in: Advanced Object Oriented Programming at Galway Institute of Technology. 

I was required to build a program that could read in a input Text File which would then build a word cloud from that file and exclude any words that were represented by an IgnoreWords folder which was hardcoded into the Project directory.